improved the UI, and moved chat from routes to app.js to improve performance
