+++ app.js
@@ -85,7 +85,7 @@ app.use('/',manageClass);
 var teachers = require('./routes/teachers.js');
 var index = require('./routes/index.js');
 var teacher_dashboard = require('./routes/teacher_dashboard.js');
-var chats = require('./routes/chatroom');
+//var chats = require('./routes/chatroom');
 var students = require('./routes/student.js');
 
 // if there are any pages that start after localhost:8080/ then route them to index
@@ -99,7 +99,7 @@ app.use('/teacher',teachers);
 app.use('/teacher_d',teacher_dashboard);
 
 // -----For the Chat
-app.use('/chat', chats);
+//app.use('/chat', chats);
 
 // ------- Student -----
 
@@ -128,7 +128,85 @@ app.use(function(err, req, res, next) {
 
 // ---- Do not remove this commented code -- Momal
 
+
+// ------------  Adding Chat here to resolve issues.
+// ----------------------------------------------------------------------------------------------
+
+
+var connection = mysql.createConnection({
+    host     : 'localhost',
+    user     : 'root',
+    password : 'root',
+    database : 'ASSESS_EASY'
+});
+
+
+
+
+var users = [];
+var connections = [];
+var userid;
+var sqlgetemail = "select email from users where userid =?";
+var insertintodb = "INSERT INTO status (s_text, classid,userid ) VALUES (?,?,?)";
+var name = "";
+var room = "";
+var groupid;
+
+app.get('/chat/:id', function(req, res) {
+    groupid = req.params.id;
+    userid = req.session.passport.user.user_id;
+    console.log('here'+ req.session.passport.user.user_id);
+    console.log('userid: ' + req.session.passport.user.user_id);
+    connection.query(sqlgetemail, userid, function (err, result) {
+        console.log('userid: ' + userid);
+        if (err) throw err;
+        else
+        {
+            name = result[0].email;
+            res.render('chatroom/chatroom');
+        }
+    });
+
+});
+
+
+
+io.on('connection', function (socket) {
+    connections.push(socket);
+    socket.username = name;
+    //console.log(socket.username);
+    console.log('from chat: ' + userid);
+    room = groupid;
+    socket.room = room;
+    socket.join(room);
+    users.push(socket.username);
+    updateUsernames();
+    socket.broadcast.to(socket.room).emit('updatechat',{ msg: socket.username + ' has joined the chat'});
+
+
+    //Disconnect
+    socket.on('disconnect', function (data) {
+        users.splice(users.indexOf(socket.username), 1);
+        updateUsernames();
+        connections.splice(connections.indexOf(socket, 1));
+       // console.log('Disconnected: %s sockets connected', connections.length);
+    });
+
+
+    //Send Message
+    socket.on('send message', function (data) {
+        //    add_message(data[0], data[1]);
+       // console.log(data)
+        // io.sockets.emit('new message', {msg: data, user:socket.username});
+        io.sockets.in(socket.room).emit('new message', {msg: data, user: socket.username});
+    });
+
+
+    function updateUsernames() {
+        io.sockets.emit('get users', users);
+    }
+});
+
 server.listen(process.env.PORT || 3000);
 console.log("Server is running ... ");
-
-module.exports = app;
+//module.exports = app;
