+++ bin/www
@@ -93,6 +93,7 @@ function onListening() {
     debug('Listening on ' + bind);
 }
 
+/*
 
 var sock = require('../routes/chats');
-sock(io);
+sock(io);*/
