--- views/chatroom/chats.ejs
@@ -1,37 +1,0 @@
-
-<!--
-<div id="userFormArea" class="row">
-    <div class="col-md-12" >
-        <form id="userForm">
-            <div class="form-group" >
-                <label>Enter Username</label>
-                <input id="username" class="form-control"/>
-                <br/>
-                <input type="submit" class="btn btn-success" value="Login"/>
-            </div>
-        </form>
-
-    </div>
-</div>
--->
-
-
-<div id="messageArea" class="row">
-    <div class="col-md-4" >
-        <div class="well">
-            <h3>Online Users</h3>
-            <ul class="list-group" id="users"></ul>
-        </div>
-    </div>
-    <div class="col-md-8" >
-        <div class="chat" id="chat"></div>
-        <form id="messageForm">
-            <div class="form-group" >
-                <label>Enter Message</label>
-                <textarea id="message" class="form-control"></textarea>
-                <br/>
-                <input type="submit" class="btn btn-success" value="Send Message"/>
-            </div>
-        </form>
-    </div>
-</div>
