+++ views/student/alltests.ejs
@@ -33,12 +33,17 @@
     <% include ../side_menu.ejs %>
 
     <div id="page-wrapper">
+        <div class="row">
+            <div class="col-lg-12">
+                <h2 class="page-header">Your Assessments</h2>
+            </div>
+        </div>
         <div class="container">
             <div class="row">
                 <div class="col col-lg-11">
-                    <h2>Your Assignments</h2>
-                    <table style="margin-top: 20px;border:none;padding: 130px;border-color:#0000" class="table table-dark">
-                        <thead style="background-color: #222d5a;color: white;border-color:#0000" class="thead-dark">
+                    <div style="margin-left: -11px" class="panel panel-default">
+                    <table width="100%" class="table table-striped table-bordered" id="dataTables-example">
+                        <thead>
                         <tr>
                             <th style="border:none;padding: 15px;">Assessment</th>
                             <th style="border:none;padding: 15px;">Total Marks</th>
@@ -59,13 +64,15 @@
                                 <% } else{ %>
                                 <span style="background-color: #a94442" class="badge badge-danger"><%= result.Due %></span>
                                 <% } %> </td>
-                            <td style="border:none;padding: 15px;"><a href="givetest/<%= result.id %>">Give Test</a></td>
+                            <td style="border:none;padding: 15px;"><a class="btn btn-success btn-sm" href="givetest/<%= result.id %>">Give Test</a></td>
 
                         </tr>
                         <% }) %>
 
                         </tbody>
                     </table>
+
+                    </div>
                 </div>
             </div>
         </div>
