+++ views/student/mcqtest.ejs
@@ -56,6 +56,7 @@
                                     </tbody>
                                 </table>
                                 <button type="submit" class="btn btn-success" >Submit</button>
+                                    <button type="button" onclick="goback()" class="btn btn-danger" >Cancel</button>
                                 </div>
                             </form>
                         </div>
