+++ views/student/results.ejs
@@ -23,13 +23,19 @@
     <% include ../side_menu.ejs %>
 
     <div id="page-wrapper">
+        <div class="row">
+            <div class="col-lg-12">
+                <h2 class="page-header">Your Results</h2>
+            </div>
+        </div>
         <div class="container">
             <div class="row">
-            <div class="col col-lg-11">
-                <h1>Student Results</h1>
-<br>
-            <table style="border:none;padding: 130px;border-color:#0000" class="table table-dark">
-                <thead style="background-color: black;color: white;border-color:#0000" class="thead-dark">
+                <div class="col col-lg-11">
+                    <div style="margin-left: -11px" class="panel panel-default">
+
+
+                        <table width="100%" class="table table-striped table-bordered" id="dataTables-example">
+                <thead>
                 <tr>
                     <th style="border:none;padding: 15px;">Assessment</th>
                     <th style="border:none;padding: 15px;">Total Marks</th>
@@ -59,6 +65,8 @@
     </div>
         </div>
     </div>
+</div>
+</div>
 </div>
 </body>
 </html>
