+++ views/teacher/viewsubmissions.ejs
@@ -32,43 +32,49 @@
     <% include ../side_menu.ejs %>
 
     <div id="page-wrapper">
+        <div class="row">
+            <div class="col-lg-12">
+                <h2 class="page-header">Submitted Assessments</h2>
+            </div>
+        </div>
         <div class="container">
             <div class="row">
                 <div style="margin-top: 20px" class="col col-lg-11">
-                    <h4>Submitted ... </h4>
-                    <table style="margin-top: 20px;border:none;border-color:#0000;border-radius: 5px" class="table table-dark">
-                        <thead style="background-color: #222d5a;color: white;border-color:#0000;border-radius: 5px" class="thead-dark">
+                    <div style="margin-left: -11px" class="panel panel-default">
+                        <table width="100%" class="table table-striped table-bordered" id="dataTables-example">
+                            <thead>
                         <tr>
-                            <th style="border:none;padding: 5px;">Student Name</th>
-                            <th style="border:none;padding: 5px;">Obtained Marks</th>
-                            <th style="border:none;padding: 5px;" ></th>
-                            <th style="border:none;padding: 5px;" ></th>
-                            <th style="border:none;padding: 5px;" ></th>
+                            <th style="border:none;padding: 15px;">Student Name</th>
+                            <th style="border:none;padding: 15px;">Obtained Marks</th>
+                            <th style="border:none;padding: 15px;" ></th>
+                            <th style="border:none;padding: 15px;" ></th>
+                            <th style="border:none;padding: 15px;" ></th>
                         </tr>
                         </thead>
                         <tbody>
                         <% result.forEach((result) => { %>
                         <tr style="border-color: transparent">
-                            <td style="border:none;padding: 5px;"><%= result.uName %></td>
-                            <td style="border:none;padding: 5px;"><%= result.Obtained %></td>
-                            <td style="border:none;padding: 5px;text-align: center"><% if(result.results.toString() == 'Added'){ %>
+                            <td style="border:none;padding: 15px;"><%= result.uName %></td>
+                            <td style="border:none;padding: 15px;"><%= result.Obtained %></td>
+                            <td style="border:none;padding: 15px;text-align: center"><% if(result.results.toString() == 'Added'){ %>
                             <span style="background-color: #3c763d" class="badge badge-success"><%= result.results %></span>
                             <% } else{ %>
                             <span style="background-color: #a94442" class="badge badge-danger"><%= result.results %></span>
                             <% } %> </td>
-                            <td style="border:none;padding: 5px;"></td>
-                            <td style="border:none;padding: 5px;text-align: center"><% if(result.results.toString() == 'Added'){ %>
+                            <td style="border:none;padding: 15px;"></td>
+                            <td style="border:none;padding: 15px;text-align: center"><% if(result.results.toString() == 'Added'){ %>
 
-                                    <a  style="color: darkslategray" onclick="return false;" >View Submissions</a>
+                                    <a class="btn btn-success btn-sm"  style="background-color:#868e96;border-color: #868e96" onclick="return false;" >View Submissions</a>
                                 <% } else{ %>
 
-                                    <a href="../studentanswers/<%= result.userID %>/<%= result.assessmentID %>">View Submissions</a>
+                                    <a class="btn btn-success btn-sm" href="../studentanswers/<%= result.userID %>/<%= result.assessmentID %>">View Submissions</a>
                                 <% } %> </td>
 
                         </tr>
                         <% }) %>
                         </tbody>
                     </table>
+                </div>
                 </div>
             </div>
         </div>
